Add GetUser AJAX Function
I removed getEmail from the chain of events because underscore.js using
map will continue going down the hole until 1 object is returned. In
this case, the last object we want to return is the PR. To get the
email, another AJAX request was created to pull the user object. This
getUser object is called from the addTableLineCommit function. This
will search the user to see if an email has been exposed on their
profile